CancelRequests - Allow multiple cancel request writers in a single job
Add the ability to require multiple cancel streams in a single job. Requiring is done as you would expect, just target the different streams on the JobConfig. Fulfilling the writer/instance is done by passing the stream instance into the `Fulfill` method.

The access wrappers now represent access to the cancel type not a specific cancel stream instance. This works because all of the cancel streams share the same SharedWrite Pending collection so getting write access for one stream grants write access for all. The instance passed through `Fulfill` will get safety checked in the access wrapper and delivered back to be used for writer creation.

Behaviour:
 - One cancel request stream required: `jobData.Fulfill` behaves as normal, no explicit instance required.
 - N cancel request stream required: `jobData.Fulfill` must be provided the instance for each stream to fulfill otherwise the safety system will throw an exception.

Note: This is a workaround implementation. #224 will make this cleaner.

This approach can be applied to all write wide (shard write) types. Support will be added in an upcoming commit.